requirements.txt
Django
south
django-mptt
django-reversion
django-tinymce
django-model-utils
sorl-thumbnail
django-pjax
django-extensions
django-widget-tweaks
django-extensions
ipython
django-babylon>=0.1.2
